% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasibinom.R
\name{dquasibinom}
\alias{dquasibinom}
\title{Quasi-Binomial Distribution}
\usage{
dquasibinom(k, size, prob, phi)
}
\arguments{
\item{k}{vector of quantiles.}

\item{size}{number of trials (greater than 0).}

\item{prob}{probability of success on each trial only when phi = 0.}

\item{phi}{dispersion parameter. Distribution is equivalent to a binomial when phi is 0.
Phi must be <= min(prob/size, 1 - prob/size) for a proper density function.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function for quasi-binomial distribution.

`dquasibinom`, `pquasibinom`, and `qquasibinom` give the density, cumulative density,
and quantile distribution functions.
}
\details{
p(x) =  	\choose{\text{size},k} \cdot \text{prob} \cdot
         (\text{prob} + x\cdot\phi)^(x-1) \cdot (1-\text{prob}-x\cdot\phi)^(\text{size}-x)

The function `quasibinom_FindParameters` uses optim to find the parameter
combination to match a given expectation and variance. Answer should not be
interpreted as exact, so proceed with caution.
}
\examples{

dquasibinom(1:30, 30, 0.5, 0.5/30)
pquasibinom(1:30, 30, 0.5, 0.5/30)
qquasibinom(0.5, 30, 0.5, 0.5/30)
}
