% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasibinom_FindParameters.R
\name{quasibinom_FindParameters}
\alias{quasibinom_FindParameters}
\title{Find Quasi-Binomial Distribution Parameters from Moments}
\usage{
quasibinom_FindParameters(size, mean, var)
}
\arguments{
\item{size}{number of trials (greater than 0).}

\item{mean}{desired mean of a quasibinomial distribution.}

\item{var}{desired variance of a quasibinomial distribution.}
}
\value{
A numeric vector. `prob` is probability of success on each trial only if phi = 0.
`phi` is the dispersion parameter. Distribution is equivalent to a binomial when phi is 0.
}
\description{
Find the parameters `prob` and `phi` for the quasibinomial
given a desired expectation, variance, and size.
}
\details{
p(x) =  	\choose{\text{size},k} \cdot \text{prob} \cdot
         (\text{prob} + x\cdot\phi)^(x-1) \cdot (1-\text{prob}-x\cdot\phi)^(\text{size}-x)

Phi must be <= min(prob/size, 1 - prob/size) for a proper density function.
The function `quasibinom_FindParameters` uses optim to find the parameter
combination to match a given expectation and variance. Answer should not be
interpreted as exact, so proceed with caution.
}
\examples{

Equasibinom(30, 0.5, 0.1/30)
Vquasibinom(30, 0.5, 0.1/30)
quasibinom_FindParameters(30, 16.6, 9.07)
}
